+++ example/lib/game.dart
@@ -1,11 +1,14 @@
-// ignore_for_file: public_member_api_docs
+// ignore_for_file: public_member_api_docs, avoid_print
 
 import "package:flame/game.dart";
+import "package:flame/input.dart";
 import "package:flame_steamworks/flame_steamworks.dart";
 import "package:flutter/material.dart";
+import "package:flutter/services.dart";
 import "package:steamworks/steamworks.dart";
 
-class GameInstance extends FlameGame with FPSCounter, SteamGame {
+class GameInstance extends FlameGame
+    with FPSCounter, KeyboardEvents, HasSteamClient {
   TextPaint textPaint = TextPaint(
     style: const TextStyle(
       fontSize: 18.0,
@@ -27,6 +30,38 @@ class GameInstance extends FlameGame with FPSCounter, SteamGame {
 
     textPaint.render(canvas, "fps: ${fps()}", Vector2(5, 20));
 
-    textPaint.render(canvas, "# of events received: $count", Vector2(5, 40));
+    textPaint.render(
+      canvas,
+      "# of chat events received: $count",
+      Vector2(5, 40),
+    );
+  }
+
+  @override
+  KeyEventResult onKeyEvent(
+    RawKeyEvent event,
+    Set<LogicalKeyboardKey> keysPressed,
+  ) {
+    bool isKeyDown = event is RawKeyDownEvent;
+
+    bool isSpace = keysPressed.contains(LogicalKeyboardKey.space);
+
+    if (isSpace && isKeyDown) {
+      CSteamId steamId = steamClient.steamUser.getSteamId();
+      SteamApiCall callId =
+          steamClient.steamUserStats.requestUserStats(steamId);
+
+      registerCallResult<UserStatsReceived>(
+        callId,
+        (ptrUserStatus, hasFailed) {
+          print("User stats received");
+          print("GameId: ${ptrUserStatus.gameId}");
+          print("SteamIdUser: ${ptrUserStatus.steamIdUser}");
+        },
+      );
+
+      return KeyEventResult.handled;
+    }
+    return KeyEventResult.ignored;
   }
 }