+++ lib/flame_steamworks.dart
@@ -1,4 +1,4 @@
 library flame_steamworks;
 
-export "src/steam_game.dart";
-export "src/steam_game_server.dart";
+export "src/has_steam_client.dart";
+export "src/has_steam_server.dart";