+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+## 0.2.0
+
+- Changed the names of `SteamGame`, `SteamGameServer` mixins to `HasSteamClient`, `HasSteamServer`, respectively, in order to comply with flame naming conventions
+- Update readme and example
+
 ## 0.1.1
+
 - Update readme
 
 ## 0.1.0
+
 - Initial release